Move logic for custom endpointslices to new module
This commit relocates the logic responsible for generating custom
endpointslices to a dedicated module.
The main `commatrix` logic now invokes `customEndpointSlices` to
generate services that aren't natively represented by endpoint slices.

The module utilizes the "ss" library to extract listening ports on
the nodes' machines. The `FilterPorts` function is employed to retrieve
only known port numbers or service names. Subsequently, endpoint slices
are created on the cluster with the temporary prefix "commatrix-test-"
to indicate their custom nature.

The list of known ports resides in "internal/ss/ports.go" and is sourced
from the OpenShift enhancements repository:

https://github.com/openshift/enhancements/blob/master/dev-guide/host-port-registry.md
