+++ internal/ss/ports.go
@@ -1,7 +1,7 @@
 package ss
 
 var (
-	tcpHostPorts = []string{"80", "443", "1936", "2041", "2379", "2380", "5050", "6080", "6180", "6181",
+	knownTCPPorts = []string{"80", "443", "1936", "2041", "2379", "2380", "5050", "6080", "6180", "6181",
 		"6183", "6385", "6388", "6443", "9001", "9099", "9100", "9101", "9101", "9102", "9102", "9103", "9105",
 		"9106", "9107", "9108", "9120", "9121", "9122", "9191", "9192", "9193", "9194", "9200", "9201", "9202",
 		"9203", "9204", "9205", "9206", "9207", "9208", "9209", "9210", "9211", "9212", "9213", "9214", "9215",
@@ -11,5 +11,21 @@ var (
 		"22624", "60000", "4180", "8797", "9259", "9260", "9443", "9977", "10248", "10300", "10301", "10302",
 		"10303", "11443", "29100", "29101", "29102", "29103", "29105", "29108", "29150", "29151", "29445", "39300", "39301"}
 
-	udpHostPorts = []string{"500", "4500", "4789", "6081", "9122"}
+	knownUDPPorts = []string{"500", "4500", "4789", "6081", "9122"}
 )
+
+func getKnownTCPPorts() map[string]bool {
+	tcpPorts := make(map[string]bool)
+	for _, port := range knownTCPPorts {
+		tcpPorts[port] = true
+	}
+	return tcpPorts
+}
+
+func getKnownUDPorts() map[string]bool {
+	udpPorts := make(map[string]bool)
+	for _, port := range knownUDPPorts {
+		udpPorts[port] = true
+	}
+	return udpPorts
+}
